+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Cycle delay helpers to `Signal` API (`reg_next`, `reg_next_with_default`)
 
 ## [0.1.16] - 2020-12-13
 ### Fixed